Extract out `completions_from_string()`
