+++ crates/ark/src/lsp/completions/sources/unique/file_path.rs
@@ -20,30 +20,10 @@ use crate::lsp::completions::completion_item::completion_item_from_direntry;
 use crate::lsp::completions::sources::utils::set_sort_text_by_words_first;
 use crate::lsp::document_context::DocumentContext;
 
-pub fn completions_from_file_path(
-    context: &DocumentContext,
-) -> Result<Option<Vec<CompletionItem>>> {
+pub(super) fn completions_from_file_path(context: &DocumentContext) -> Result<Vec<CompletionItem>> {
     log::info!("completions_from_file_path()");
 
-    let node = context.node;
-
-    if node.kind() != "string" {
-        return Ok(None);
-    }
-
-    // Must actually be "inside" the string, so these places don't count, even
-    // though they are detected as part of the string nodes `|""|`
-    if node.start_position() == context.point || node.end_position() == context.point {
-        return Ok(None);
-    }
-
     let mut completions: Vec<CompletionItem> = vec![];
-
-    // Return empty set if we are here due to a trigger character like `$`.
-    // See posit-dev/positron#1884.
-    if context.trigger.is_some() {
-        return Ok(Some(completions));
-    }
 
     // Get the contents of the string token.
     //
@@ -77,6 +57,7 @@ pub fn completions_from_file_path(
     // look for files in this directory
     log::info!("Reading directory: {}", path.display());
     let entries = std::fs::read_dir(path)?;
+
     for entry in entries.into_iter() {
         let entry = unwrap!(entry, Err(error) => {
             log::error!("{}", error);
@@ -95,73 +76,5 @@ pub fn completions_from_file_path(
     // the sort list (like those starting with `.`)
     set_sort_text_by_words_first(&mut completions);
 
-    Ok(Some(completions))
-}
-
-#[cfg(test)]
-mod tests {
-    use harp::assert_match;
-    use tree_sitter::Point;
-
-    use crate::lsp::completions::sources::completions_from_unique_sources;
-    use crate::lsp::completions::sources::unique::file_path::completions_from_file_path;
-    use crate::lsp::document_context::DocumentContext;
-    use crate::lsp::documents::Document;
-    use crate::test::r_test;
-
-    #[test]
-    fn test_file_path_outside_quotes() {
-        r_test(|| {
-            // Before or after the `''`, i.e. `|''` or `''|`.
-            // Still considered part of the string node.
-            let point = Point { row: 0, column: 0 };
-            let document = Document::new("''");
-            let context = DocumentContext::new(&document, point, None);
-
-            assert_eq!(context.node.kind(), "string");
-            assert_eq!(completions_from_file_path(&context).unwrap(), None);
-        })
-    }
-
-    #[test]
-    fn test_file_path_not_string() {
-        r_test(|| {
-            let point = Point { row: 0, column: 0 };
-            let document = Document::new("foo");
-            let context = DocumentContext::new(&document, point, None);
-
-            assert_eq!(context.node.kind(), "identifier");
-            assert_eq!(completions_from_file_path(&context).unwrap(), None);
-        })
-    }
-
-    #[test]
-    fn test_file_path_trigger() {
-        r_test(|| {
-            // Before or after the `''`, i.e. `|''` or `''|`.
-            // Still considered part of the string node.
-            let point = Point { row: 0, column: 2 };
-
-            // Assume home directory is not empty
-            let document = Document::new("'~/'");
-
-            // `None` trigger -> Return file completions
-            let context = DocumentContext::new(&document, point, None);
-            assert_match!(
-                completions_from_file_path(&context).unwrap(),
-                Some(items) => {
-                    assert!(items.len() > 0)
-                }
-            );
-
-            // `Some` trigger -> Should return empty completion set
-            let context = DocumentContext::new(&document, point, Some(String::from("$")));
-            let res = completions_from_file_path(&context).unwrap();
-            assert_match!(res, Some(items) => { assert!(items.len() == 0) });
-
-            // Check one level up too
-            let res = completions_from_unique_sources(&context).unwrap();
-            assert_match!(res, Some(items) => { assert!(items.len() == 0) });
-        })
-    }
+    Ok(completions)
 }
